Renamed all keyword constants by removing the "JML_" prefix.  This caused some name clashes, so I also renamed PRE and POST-related constants to PRECONDITION and POSTCONDITION, respectively.
Detected an infinite loop (out of memory) bug in handling of ignored
keywords and clauses.  This is caused by the fact that different kinds
of clauses require different skipping techniques.  I've refined the
fall-through case in the pragma parser to deal with this, but we still
are not handling "abrupt_behavior" correctly.  Unfortunately, I don't
know how this clause is defined, so I cannot yet write the right
support for it.  I'm searching through the JML2 docs, but so far I
have not found the answer I'm looking for.