MVA.pvs .Changed source and target in the definition of OneAction .Added the predicate valid_MVA. It is needed to contrain the possible values of the components of a MVA. In the process I defined valid_Store and valid_PSt .Assuming the argument is a valid MVA, simplified the definition of deterministic. The definition was   deterministic(a : (valid)) : boolean =     FORALL(q : CP, e : Event, stA : Store, stP : PSt,            p : Guard, t : Transition) :       cps(a)(q) IMPLIES       evs(a)(e) IMPLIES       ts(a)(t) IMPLIES       scp(t) = q IMPLIES       event(t) = e IMPLIES       guard(t) = p IMPLIES       p(stA, stP) IMPLIES       FORALL (t1 : Transition) :         scp(t1)(q) IMPLIES         event(t1)(e) IMPLIES         guard(t1) /= p IMPLIES         NOT guard(t1)(stA, stP)
Since now we assume that the MVA is valid we can take out the predicates
      cps(a)(q) IMPLIES
      evs(a)(e) IMPLIES
We also need to relate stA and stP with the automaton. We must say that they are defined for each varible as we did in the definition of valid_MVA. Besides, this definition is wrong because a MVA with the transitions (s1, e1, p1, a1, s2) and (s1, e1, p1, a2, s3) is not deterministic but it satisfies the previous definition.

I defined it as:
  deterministic(a : (valid_MVA)) : boolean =
    FORALL (t1, t2 : Transition, stA : Store, stP : PSt) :
      ts(a)(t1) IMPLIES ts(a)(t2) IMPLIES
      valid_Store(stA, a) IMPLIES valid_PSt(stP, a) IMPLIES
      scp(t2) = scp(t1) IMPLIES
      event(t2) = event(t1) IMPLIES
      guard(t1)(stA, stP) IMPLIES
      guard(t2)(stA, stP) IMPLIES
      t1 = t2

.Proved the lemma MC_TCC2_subset1_is_finite used in the proof completeTrans_TCC1.


		Semantincs.pvs
Definied
  lift_PSt(p : PSt) : PState =
    (# ex := bottom, st := p #)

  CONVERSION lift_PSt
to be able to use the function defined PState on a PSt